requirements.txt
Flask==2.2.3
Flask-RESTful==0.3.9
Flask-SQLAlchemy==3.0.3
Flask-JWT-Extended==4.5.2
Flask-Cors==3.0.10
SQLAlchemy==2.0.4
pymysql==1.0.3
cryptography==39.0.1
python-dotenv==1.0.0
bcrypt==4.0.1
marshmallow==3.19.0
PyJWT==2.6.0
Werkzeug==2.2.3
